Fix chargePeriodStart vs periodStart in trans.
When we were working on the spike in [PR #81](https://github.com/DEFRA/sroc-charging-module-api/pull/81) we had the intent of fixing some of the transaction field names to make them simpler and remove unnecessary prefixes.

We've decided to hold off on those changes and focus on ensuring we have things behaving as they currently do first. One of those changes that have managed to be implemented is referring to the `chargePeriodStart` field as `periodStart`.

To make things consistent with how the table has now been set up this small change rolls back the use of `periodStart` instead of `chargePeriodStart`.