Retain rules service response in translator
Currently, the [charging-module-api](https://github.com/defra/charging-module-api) retains a formatted form of the response received from the charging module. This makes a lot of sense from an auditing context. Being able to show the calculation response if questioned about the charge value assigned is handy.

We want to do the same in this project so this change adds support for it. We do this by updating the `RulesServiceTranslator` to add it as a property. We know the properties of the translator will be assigned to a `TransactionTranslator` instance. That instance will then form the basis of what gets persisted to the db.

The one change we're making to the CHA is we are not doing any reformatting of the response. If the intent is to hold this for auditing purposes then we really should record exactly what we received, not an altered version of it.