Update charge translator regards ruleset
The `CalculateChargeService` has to handle requests coming direct via the calculate charge endpoint as well as indirectly via the create transaction endpoint.

When we create a transaction we'll have already set the rulset property. So we've updated the comments to reflect this. We have also added some tests to cover the defaulting behaviour of the calculate charge translator.