Add default ruleset property to Trans. translator (#115)
One of the changes we have made from the [charging-module-api](https://github.com/defra/charging-module-api) is that instead of using a boolean field to tell us whether something is related to `presroc` or not, we instead hold the ruleset name.

We believe this will give us much more flexibility going forward whilst still being able to tell us clearly which ruleset a transaction is related to. It also means we can use this value directly, for example, when trying to determine which rules service endpoints to hit.

As we only have `presroc` at the moment, this change adds it as a field to the `TransactionTranslator` and handles defaulting the value.

* Note

The `CalculateChargeService` has to handle requests coming direct via the calculate charge endpoint as well as indirectly via the create transaction endpoint.

When we create a transaction we'll have already set the rulset property. So we've updated the comments to reflect this. We have also added some tests to cover the defaulting behaviour of the calculate charge translator.