Fix transaction translator test
Spotted after adding bill run ID as a required field in the transaction translators Joi schema. This test should have failed but instead passed. After some investigation found it didn't seem to be instantiating an instance at all.

After some headscratching realised we don't need any fancy syntax to confirm we don't get an error. If we can instantiate an instance using the data provided and not get an error back then we will have confirmed the behaviour expected of the test.