Update Trans. Translator to expect bill run id (#116)
* Update Trans. Translator to expect bill run id

We know we'll have a bill run ID because requests to create a transaction will come via `POST /v2/bill-runs/{id}/transactions`. We also will be persisting this to the `transactions` table when we create the transaction.

But the current design does not inject the bill run ID into the translator. Instead it assumes this will be added separately when the transaction is persisted.

The issue is this forces us to generate an object to persist made up of the transaction translator, bill run ID, regime ID etc. It's not a major problem but we believe a cleaner, clearer design is to think of a request not just as the body recieved. Instead to consider all information sent

- the request body
- the regime it is for
- the authorised system making the request
- the bill run it's related too (if relevant)

So this change is the first step in implementing this concept. That translators receive and interpret _all_ request data, not just the body.

* Fix transaction translator test

Spotted after adding bill run ID as a required field in the transaction translators Joi schema. This test should have failed but instead passed. After some investigation found it didn't seem to be instantiating an instance at all.

After some headscratching realised we don't need any fancy syntax to confirm we don't get an error. If we can instantiate an instance using the data provided and not get an error back then we will have confirmed the behaviour expected of the test.

* Update transaction translator to require billRunId