Manage values to persist in bill run translator
We're trying to simplify and be consistent with the creation process for our records. In support of this we're trying to have all values that are to be persisted determined in the translator. We then can pass the translator as is to the [Objection insert query](https://vincit.github.io/objection.js/guide/query-examples.html#insert-queries).

This change updates the the BillRun translator and the create bill run service to follow this pattern.