Fix bill run translator test
We recently found the style of expect() we were runnning was not actually initialising an instance of BillRunTranslator. So the test was passing by luck and not checking the behaviour.

By switching to this form we have confirm we do instantiate an instance and that as long as the data is valid we don't get an error.