Simplify bill run insert using spread operator
We can can't just pass in the bill run translator instance as is because it we also need to set the bill run number and that comes from `NextBillRunNumberService`. But we can simplify the insert by using the spread operator on the bill run translator instance rather than specifying each field.