Manage values to persist in bill run translator (#119)
We're trying to simplify and be consistent with the creation process for our records. In support of this, we're trying to have all values that are to be persisted determined in the translator. We then can pass the translator as is to the [Objection insert query](https://vincit.github.io/objection.js/guide/query-examples.html#insert-queries).

This change updates the BillRun translator and the create bill run service to follow this pattern.

* Note - Includes a fix for a bill run translator test

We recently found the style of `expect()` we were running was not actually initialising an instance of BillRunTranslator. So the test was passing by luck and not checking the behaviour.

By switching to this form we have confirmed we do instantiate an instance and that as long as the data is valid we don't get an error.