Simplify auth. system creation in service
Similar to the changes made in [Manage values to persist in bill run translator](https://github.com/DEFRA/sroc-charging-module-api/pull/119) we are trying to simplify and be consistent with the creation process of records. We want to be able to pass in translator instances to [Objection insert queries](https://vincit.github.io/objection.js/guide/query-examples.html#insert-queries) rather than spell out each field that needs updating.

We've made the change for when creating bill runs. At the moment the only other `create` is for authorised systems. This updates the `CreateAuthorisedSystemService` and associated `AuthorisedSystemTranslator` to support this new convention.