Update create auth. system service
With the `AuthorisedSystemTranslator` all setup as needed we can pass it into the insert query directly rather then listing out each field to update. We have to combine it with `regimes` because that data is coming from elsewhere.

Finally, we need access to the validated `authorisations` but we can no longer access them as a property of `AuthorisedSystemTranslator` directly. But thanks to a previous change which exposes the vaidated data object we can still access them.