Simplify auth. system creation in service (#123)
Similar to the changes made in [Manage values to persist in bill run translator](https://github.com/DEFRA/sroc-charging-module-api/pull/119) we are trying to simplify and be consistent with the creation process of records. We want to be able to pass in translator instances to [Objection insert queries](https://vincit.github.io/objection.js/guide/query-examples.html#insert-queries) rather than spell out each field that needs updating.

We've made the change for when creating bill runs. At the moment the only other `create` is for authorised systems. This updates the `CreateAuthorisedSystemService` and associated `AuthorisedSystemTranslator` to support this new convention.

* Note - Includes a refactor of the auth. system translator tests

These were the first translator tests we added. Since then we have implemented a new pattern for translator tests which includes a few fixes. This first change updates the authorised system translator tests to match the existing pattern and convention.

* Note - Need to remove authorisations from the translations

If we include authorisations in the translations, when we come to use the translator instance in the insert query it errors because there is no `authorisations` field in the `authorised_systems` table.

So we still validate it, but we don't translate it. This is also why we now access the validated authorisations via the translator's `validatedData` getter.