Simplify create transaction call
By just getting the rules service result and assigning it to our translator we have an instance we can pass as is to the insert query rather than having to detail each field.

As part of this we have moved the transaction translator to its own method to make it clearer what's going on.