Fix calculate charge translator test
We've seen this in a few translator tests. We assumed it passing meant it was correctly testing the behaviour of the translator. However, having added `regime` as a required property it should be failing but its not. Previous investigations found that it's because the translator never actually gets instantiated in the test. Using this new format ensures it does and that we are correctly testing the expected behaviour.