Expose regime in the translations
We didn't do this initially because there is no `regime` field in the transactions table and we thought this might cause problems. However, it's not the `CalculateChargeTranslator` we bind to a `TransactionTranslator` but a `RulesServiceTranslator`.

We need the regime to be exposed for the `RulesServicePresenter` to see it. So as our concern is incorrect we go ahead and expose it in the translations.