Update Calculate Charge Translator expect Regime slug (#125)
> Part of the translator 'request' concept update kicked off by [PR #116](https://github.com/DEFRA/sroc-charging-module-api/pull/116)

This change updates the calculate charge translator to expect a regime 'slug' as part of the data passed to it.

**Concept update**

Prior to this change, the purpose of translators was to 'translate' just the request body into properties our models and services would understand.

The update expands the concept of translators to 'translate' everything in the request. So this includes regime, authorised system, and bill run, for example. They are *not* included in the body. But they are part of the request.

* Note - Fix calculate charge translator test

We've seen this in a few translator tests. We assumed it passing meant it was correctly testing the behaviour of the translator. However, having added `regime` as a required property it should be failing but it's not. Previous investigations found that it's because the translator never actually gets instantiated in the test. Using this new format ensures it does and that we are correctly testing the expected behaviour.

* Note - Refactor calculate charge translator tests

This brings the tests in line with the pattern and conventions used in the other translator tests.