Access rules service result from charge service (#126)
We're endeavouring to keep things simple, clean and avoid duplication. There are 2 places we need to accept a request, forward the details to the rules service, and return a response

- standalone calculate charge endpoint
- create a transaction endpoint

For both the flow is pretty similar right up to the point we get the result back from the rules service and translate it into something we understand.

- The standalone calculate charge then passes it to a presenter ready for returning to the client
- The create transaction needs the result in a format it understands for persisting to the database

We have experimented with changes to the presenters and translators, and even breaking up the charge service. However, we realised we were making the solution much more complicated simply to stick to a pattern of 'one input - one output'. By providing an override flag to the `CalculateChargeService` we can have it return the result needed for each scenario.

So this change updates the service to allow the caller to determine what format of result is needed.