Fix sucFactor from rules service not translated (#127)
Spotted that when we get the result from the `RulesService` we are storing it in the `RulesServiceTranslator` as `sucFactor`. Our service understands this as `lineAttr4` on the transaction. We did have the translation listed. But we were also adding a `sucFactor` property. As can be seen in the subsequent line our convention in a translator _if_ we have to set a property is to use the translated name.

So this updates `RulesServiceTranslator` to fix this issue. The end result is when we use the translated fields in an [Objection insert query](https://vincit.github.io/objection.js/guide/query-examples.html#insert-queries) the fields will match what is in the table.