Add create transaction service (#107)
https://trello.com/c/gTcQmR8M

This change adds a create transaction service to the app. The service will orchestrate taking the initial request payload and

- translating and validating it using the `TransactionTranslator`
- determining the charge value using the `CalculateChargeService`
- persisting the transaction to the database
- returning the appropriate response

A lot of the work will be based on in [PR #81](https://github.com/DEFRA/sroc-charging-module-api/pull/81) and our spike on creating transactions with summaries.

* Note - Add new transaction fixture

To support the tests for the new service this change also includes adding a new fixture. It will give us a base request to work from and is taken from the Postman files we were given on starting the project.