Add missing test coverage 4 RulesServiceTranslator
Our test coverage has been showing a couple of lines always being missed in the `RulesServiceTranslator`. These were in the `_convertToPence()` method and were to handle if the param passed in was not a number.

In the context of the `RulesServiceTranslator` this would never happen. We use this method with `chargeValue` and `sucFactor` and the validation requires that they are numbers. We'd get a validation before we ever attempted to pass in null to `_convertToPence()`.

So the first part of this fix is to remove the code that will never be used. Then whilst we were looking at `sucFactor` we spotted there was no specific test to confirm it gets converted to pence. So this change also adds a missing test for that.