Refactor add bill run transaction controller (#129)
Because of the design we're currently following our controllers are fairly simple. All work is done in services which means the handlers are generally only a couple of lines each. As such this change refactors the `PresrocAddBillRunTransactionController` to move its functionality into the existing `BillRunsController`.

This will hopefully help when it comes to understanding where endpoints are. If, for example, you are looking for something on `/v2/wrls/bill-runs` you should be able to find it in the bill runs controller. It may be a decision we need to review as we add more endpoints. But for now, it will help keep things simple.