Remove transactions controller (#130)
This was added at the start of the project as a demonstration of how we could structure our controllers across regimes. It doesn't have any functionality, and we are some way off actually implementing support for the transaction endpoints.

As we move towards this version 2 of the API being tested and put through its paces, we're taking this opportunity to remove something that is currently purely for demo purposes.