Add default 4 status in bill run tbl / Remove dupe
The staus for all bill runs when first created is set to 'initialised'. For other tables we've set a default for the field so this change adds one for bill runs.

When looking into this we also spotted that we have a duplicate alter bill runs script. So this also removes the duplicate.