Remove bill run number from transactions (#133)
This deals with 2 issues. The first is that `db/migrations/20201210145745_alter_bill_runs.js` actually alters the `transactions` table, not the `bill_runs`. So getting rid of this script solves a problem with its confusing name, especially as the content is almost exactly the same as `db/migrations/20201215095327_alter_bill_runs.js`!

The second is an issue of performance. When a transaction is added in the [charging-module-api](https://github.com/defra/charging-module-api) it will query the DB for the bill run instance in order to confirm the bill run exists. At the same time, it means it can also grab the bill run number for adding to the transaction.

In our service, we intend to do additional work around tracking totals for invoices and licences when a transaction is added. So we really need to be conscious of every query we fire. We get the bill run ID from the request and if we ensure through the DB you can't add a transaction if the bill run doesn't exist we won't need the extra query in the app to grab the bill run first.

But this also means we won't have access to the bill run number when we add the transaction. As we'll have the link between the transaction and the bill run through the `bill_run_id` there is actually no need to store the bill run number against the transaction. So this change removes the script which adds `bill_run_number` to transactions. It also solves our other problem!