Add foreign key constraint to bill run id in trans
This adds a foreign key constrainst on the `bill_run_id` field in the `tranasactions` table.

By adding the constrainst in the table it means we don't have to query for a matching bill run every time a transaction is added. We can manage by exception those times when someone attempts to add a transaction for a bill run that does not exist. We do it this way for performance reasons.