Update existing create transactions script
We've chosen to update the existing create script rather than add an alter script. This is because there is a difference between the syntax used when setting the constraint on a new column and adding it to an existing one.

Prior to the first release of a service we like to 'squash' our migration scripts so we are not then having to maintain all the migrations we generated whilst pulling together the app. Once released they become fixed and never updated.

So when we come to do this we don't won't to have to try and remember there is a difference in the syntax. So even though it will break existing installs, we're making the change in the original create script.