Add foreign key constraint to bill run id in trans (#134)
This adds a foreign key constraint on the `bill_run_id` field in the `transactions` table.

By adding the constraint in the table it means we don't have to query for a matching bill run every time a transaction is added. We can manage by exception those times when someone attempts to add a transaction for a bill run that does not exist. We do it this way for performance reasons.

* Note - Update existing create transactions script

We've chosen to update the existing create script rather than add an alter script. This is because there is a difference between the syntax used when setting the constraint on a new column and adding it to an existing one.

Prior to the first release of a service we like to 'squash' our migration scripts so we are not then having to maintain all the migrations we generated whilst pulling together the app. Once released they become fixed and never updated.

So when we come to do this we won't have to try and remember there is a difference in the syntax. Even though it will break existing installs, we're making the change in the original create script.