Alter transactions to use BigInts
https://trello.com/c/HPENPD7d

We have previously faced an issue with the differences between what a PostgreSQL integer can hold and what a JavaScript one can. The issue was spotted when attempting to generate bill runs where the total value was more than 2147483647 pence.

We fixed it for bill runs. It was highlighted at the time that _if_ the calculated charge value for a single transaction was more than £21.4 million we'd face the same problem there. We put off updating the transaction because there is DB size and performance implications to using BigInts vs Integers. But latest intelligence shows though we've never had a single transaction for more than £21.4 million, proposed changes makes it likely that we might.

So this change updates the relevant fields in the `transactions` table to use [PostgreSQL BigInts](https://www.postgresql.org/docs/13/datatype-numeric.html#DATATYPE-INT).