Add failing tests for Bill run `editable()`
Before we can reject adding a transaction to a bill run, we need a mechanism to for determining if the bill run is editable. Like `$active()` on `AuthorisedSystemModel` this change intends to follow the pattern of asking the instance. This is a preferred pattern to the alternate which is to ask for the bill run for its status, and then making the determination in the calling code.