Fix snake case fields in database (#150)
When creating the database, we mirrored the existing database's fields as closely as possible, and used objection's snakeCaseMappers to allow us to use camelCase in the code and snake_case in the database. However we found that when we referred to a field in the code such as lineAttr1, it would be mapped to line_attr1 whereas the original name in the existing database is line_attr_1. We therefore adopted line_attr1 (and header_attr1, regime_value1, etc.) as the new standard for these fields.

We recently discovered the underscoreBeforeDigits option in objection, which causes the name to be correctly mapped to line_attr_1. We therefore decided that while the codebase is still relatively small, we should enable this option and rename the fields. This is implemented in this branch.

We also discovered knexSnakeCaseMappers, which shifts responsibility of mapping names from objection to knex. This means that we can refer to fields by the camel case name throughout the code, whereas the original snakeCaseMappers has some exceptions (for example, we still need to use snake case in queries). This is also included in this branch.

Unfortunately this causes a problem with migrations; once knexSnakeCaseMappers is enabled, the underscoreBeforeDigits rule acts on everything, camel case or not. This means that an underscore is always inserted before a digit, so creating a field line_attr_1 will actually create line_attr__1. To remedy this, we create an additional knexfile.application.js to be used by the application, which reads in the basic configuration from knexfile.js and adds in knexSnakeCaseMappers. The migrations use the basic knexfile.js and therefore work as normal.