Fix missing knexfile.application.js to docker
We recently [made a change](https://github.com/DEFRA/sroc-charging-module-api/pull/150) to start using Knex's snake case mappers rather than Objection's.

To deal with migrations we needed to add a new `knexfile.application.js` file to the project. However, we neglected to update the Docker build, specifically the `.dockerignore` to ensure its included when the image is built.

This fix corrects the issue.