Update Dockerfile with ARG and ENV statements
For reference, `ARG` is only available at build time and can be set in the Dockerfile and overridden in the `docker build` command using `--build-arg VAR_NAME=foo`.

`ENV` is available both at build and run time. If there is a value you'd like to 'bake-in' to your image so its available at runtime, using a combination of `ARG` and `ENV` is a common practice.

In our case we expect the values for `GIT_COMMIT` and `DOCKER_TAG` to be provided in the build command

```
sh "docker build -t ${git_image_tag} . --build-arg GIT_COMMIT=$(git rev-parse HEAD) --build-arg DOCKER_TAG=${git_image_tag}"
```

The line `ENV GIT_COMMIT $GIT_COMMIT` then ensures the value is available as an environment variable whenever a container is started (bakes it in).

For reference, we found this blog post a useful primer on the differences between `ARG` and `ENV` and this build pattern https://vsupalov.com/docker-arg-vs-env/.