Update compose and vscode tasks.json to support
The `docker-compose.yml` and `.vscode/tasks.json` are used to support running and developing the project locally using Docker. We want our new version values available within our local docker environment so we can see the behaviour we would expect in production.

These changes

- mimic what will happen when we build the image in production
- supports the fact that locally we'll update and call `docker-compose up` many times

A constraint is we cannot call `git rev-parse HEAD` directly in the compose file. This is why we needed to update the vscode tasks file. We call the command there and create a local env var on the fly, which docker-compose will then see when `up` is called.