Remove local override support
Because it ain't working! No doubt something we are doing wrong, though our current belief is because it's something to do with the differences between `docker-compose up` and `docker-compose run`.

But it's also a nice to have. The critical bit is the running container in our AWS environments will have access to this information and that we can mimic it locally (but not panic if the commit returned doesn't match the exact local commit on our local machine!)