Add ARG and ENV commands to production build stage
Spotted we'd only made the changes to the `development` build stage. These need to be available in the `production` build stage as well. We cannot put them into the `base` stage because the fact they'll be different for each build would negate the benefit of having a base build stage.