Create Minimum Charge Adjustment service (#156)
When minimum charge is applied to an invoice, an adjustment transaction is created. These transactions are created with a specified value, whereas regular transactions go out to the rules service to calculate their value. Therefore we need a way of creating transactions with a known value, which will be handled by the CreateMinimumChargeAdjustment service introduced in this change.