Add trans. region must match bill run check (#162)
https://trello.com/c/YleRl9ts

In version 1 of the API, a check is made against the region of a transaction being added to ensure it matches the region of set against the bill run. We are not doing this check-in version 2. If a client did add a transaction with a different region to the bill run it would cause major problems upstream when the transaction file is generated and passed to SOP.

This change adds the check to the create transaction process.