Fix bill run translator
Tracked the issue down to the fact that because of our payload cleansing a payload like this

```
{
    "region": ""
}
```

reaches the controller as `{}`. For whatever reason this was not triggering the validation in the `BillRunTranslator` which is why it was allowed to procreed. The actual error was coming from the attempt to then get the next sequence number for the bill run which depends on querying the sequences table (see `NextBillRunNumberService`).

Adding `required()` to the validation resulted in the behaviour we want to see.