Add test to highlight issue
Basically, we had missed `eiucSource` in the `CalculateChargeTranslator` 😱

We are validating it fine based on what `compensationCharge` is set to. But the rules service throws a 500 and an utterly indecipherable error message if you send a payload where `compensationCharge` is true but `eiucSource` is missing.

So, we are getting a 500 from the rules service which causes us to report a 500 back to the user.