Remove new tests
It seems odd to remove a test we've just added. But ommitting the `eiucSource` translation was just a complete 'drop of the ball'. We don't have tests that the other values translate as expected. So should someone come across these tests in the future it'd be hard to see why they exist and others don't.

So to keep things consistent I'm removing them. Should we find we've missed things on other translations then we will have to revisit the idea of testing the actual translations.

But for now I'm happy this issue is resolved.