Minimum charge database changes (#165)
At present, we apply minimum charge to licences which are flagged as "new licences". However minimum charge may be more widely applied than this in future. We are therefore renaming newLicence wherever it occurs to subjectToMinimumCharge.

Future changes also mean that an invoice could potentially have a mix of transactions, some of which are subject to minimum charge and some of which aren't. Therefore we cannot guarantee that, for example, a licence with a credit value of 5000 is not subject to minimum charge; if 1000 of this is flagged as subject to minimum charge then a minimum charge adjustment would be needed to bring this portion up to 2500. We are therefore also adding subject_to_minimum_charge_credit_value and subject_to_minimum_charge_debit_value fields at the bill run, invoice and licence level in order to determine whether an invoice is truly subject to minimum charge.