Fix create bill run returns 500 if region missing (#163)
https://trello.com/c/MJ6w1HYe

When the region is missing from the create bill run payload we are returning a `500` when we should be returning a `422` with a nice error message.

This change fixes it.

* Note

Tracked the issue down to the fact that because of our payload cleansing a payload like this

```
{
    "region": ""
}
```

reaches the controller as `{}`. For whatever reason, this was not triggering the validation in the `BillRunTranslator` which is why it was allowed to proceed. The actual error was coming from the attempt to then get the next sequence number for the bill run which depends on querying the sequences table (see `NextBillRunNumberService`).

Adding `required()` to the validation resulted in the behaviour we want to see.