Fix 500 error when compensation charge set to true (#166)
https://trello.com/c/YleRl9ts
https://trello.com/c/xoHCMKqf

When adding a bill run transaction or calling the calculate charge endpoint if `compensationCharge` is true we are returning a `500` error.

This change fixes the issue.

* Note

Basically, we had missed `eiucSource` in the `CalculateChargeTranslator` 😱

We were validating it fine based on what `compensationCharge` is set to. But the rules service throws a 500 and an utterly indecipherable error message if you send a payload where `compensationCharge` is true but `eiucSource` is missing.

So, we were getting a 500 from the rules service which causes us to report a 500 back to the user. Add the missing translation and everything works! Rules service no longer returns a 500 so we don't either.