Fix Joi validations using US date format
For folks not aware, UK dates are written as either `YYYY-MM-DD` or `DD-MM-YYYY`. In the US month and day are reversed so they use `YYYY-DD-MM` or `MM-DD-YYYY`.

We use [Joi](https://github.com/sideway/joi) to validate our requests, which originated with US based contributors. So when it validates a date string it will default to a US date format.

The convention with the API is that folks have been passing period end and start dates as `DD-MMM-YYYY`. But there is nothing to stop anyone using any valid date string as long Joi can validate it.

Our problem is _if_ a client system chooses to provide these date values as `DD-MM-YYYY` Joi will convert them to US format. So, a UK date of 31-03-2021 will be seen by Joi as 3rd of the 31st month 2021!

This fixes the issue and ensures Joi defaults to a UK based date format when attempting to convert date strings.