Add tests for date formats we may expect
Currently, it's just the period start and end dates we have to parse and validate, whenever a charge needs calculating.

Understanding what was required to get the API able to accept various formats has taken some work. So, at this point we already know the answer! But we are breaking down the commits for easier review.