Fix Joi validations using US date format (#170)
For folks not aware, UK dates are written as either `YYYY-MM-DD` or `DD-MM-YYYY`. In the US month and day are reversed so they use `YYYY-DD-MM` or `MM-DD-YYYY`.

We use [Joi](https://github.com/sideway/joi) to validate our requests. What we have found is that when passed dates in certain formats it appears to use the US format rather than the UK. For example, it will see `01-04-2020` as 4 January 2020. `31-03-2020` will cause the API to return an invalid date error.

The convention with the API is that folks have been passing period end and start dates as `DD-MMM-YYYY`. But we don't want to force client systems to use only this format. We want to allow as many reasonable variations as possible.

We had hoped there would be a global 'catch-all' solution, either at the system or Joi level which would allow us to default to expecting dates in a UK format. But our understanding is that Joi relies on whatever the JavaScript [Date.parse()](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date/parse) returns

> If the validation convert option is on (enabled by default), a string or number will be converted to a Date if specified. Note that some invalid date strings will be accepted if they can be adjusted to valid dates (e.g. '2/31/2019' will be converted to '3/3/2019') by the internal JS Date.parse() implementation.
>
> https://joi.dev/api/?v=17.3.0#date

Our research concluded that

- you need to add the [joi-date extension](https://joi.dev/module/joi-date/)
- you then have to be explicit with the formats you expect

This change implements those changes to allow us to handle properly the expected date formats for `periodEnd` and `periodStart` when creating a transaction or calculating a charge.